refactor to Int and max limit upto 9

+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/js/userinfo/Points.kt
@@ -5,7 +5,7 @@ import androidx.annotation.Keep
 /** Points object for miniapp. */
 @Keep
 data class Points(
-    val standard: Long,
-    var term: Long,
-    var cash: Long
+    val standard: Int,
+    var term: Int,
+    var cash: Int
 )
